move url context provider to upper level
